requirements.txt
asgiref==3.10.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
Django==5.2.8
djangorestframework==3.15.2
idna==3.11
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
